refactor: add view to see users for admin + new comp for the table
